Move netty to a separate module
