+++ instrumentation/build.gradle.kts
@@ -40,6 +40,7 @@ dependencies{
     implementation(project(":instrumentation:jaxrs-client-2.0"))
     implementation(project(":instrumentation:java-streams"))
     implementation(project(":instrumentation:apache-httpasyncclient-4.1"))
+    implementation(project(":instrumentation:netty:netty-4.0"))
     implementation(project(":otel-extensions"))
 }
