+++ instrumentation/netty/netty-4.0/src/main/java/io/opentelemetry/javaagent/instrumentation/hypertrace/netty/v4_0/NettyInstrumentationName.java
@@ -14,15 +14,11 @@
  * limitations under the License.
  */
 
-package io.opentelemetry.javaagent.instrumentation.hypertrace.vertx.netty.server;
+package io.opentelemetry.javaagent.instrumentation.hypertrace.netty.v4_0;
 
-import io.netty.channel.CombinedChannelDuplexHandler;
-
-public class HttpServerTracingHandler
-    extends CombinedChannelDuplexHandler<
-        HttpServerRequestTracingHandler, HttpServerResponseTracingHandler> {
-
-  public HttpServerTracingHandler() {
-    super(new HttpServerRequestTracingHandler(), new HttpServerResponseTracingHandler());
-  }
+public class NettyInstrumentationName {
+  public static final String PRIMARY = "netty";
+  public static final String[] OTHER = {
+    "netty-4.0", "ht", "netty", "netty-4.0-ht",
+  };
 }
